--- src/javascripts/components/eat.js
@@ -1,4 +1,0 @@
-// import printToDom from '../helpers/printToDom';
-// import { full, healthy, unhealthy } from '../helpers/data/eatData';
-
-// export default eat;
