--- src/javascripts/helpers/domEvents.js
@@ -1,11 +1,0 @@
-// import progressUpdate from './data/progressData';
-// import { full } from './data/eatData';
-// import { fun } from './data/playData';
-// import { strength } from './data/fightData';
-// import { energy } from './data/sleepData';
-
-// const domEvents = () => {
-//   progressUpdate(full, fun, strength, energy);
-// };
-
-// export default domEvents;
